requirements.txt
Flask==1.1.1
PyJWT==1.7.1
Werkzeug==0.16.0
fort==0.0.7
psycopg2==2.8.4
requests==2.22.0
waitress==1.3.1
